+++ controlplane/src/core/bufservices/api-key/deleteAPIKey.ts
@@ -1,4 +1,3 @@
-import { PlainMessage } from '@bufbuild/protobuf';
 import { HandlerContext } from '@connectrpc/connect';
 import { EnumStatusCode } from '@wundergraph/cosmo-connect/dist/common/common_pb';
 import { DeleteAPIKeyRequest, DeleteAPIKeyResponse } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_pb';
@@ -9,6 +8,7 @@ import { enrichLogger, getLogger, handleError } from '../../util.js';
 import { UnauthorizedError } from '../../errors/errors.js';
 import { OrganizationGroupRepository } from '../../repositories/OrganizationGroupRepository.js';
 import { RBACEvaluator } from '../../services/RBACEvaluator.js';
+import type { PlainMessage } from '../../../types/index.js';
 
 export function deleteAPIKey(
   opts: RouterOptions,