+++ controlplane/src/core/bufservices/federated-graph/getRouters.ts
@@ -1,9 +1,8 @@
-import { PlainMessage } from '@bufbuild/protobuf';
 import { HandlerContext } from '@connectrpc/connect';
 import { EnumStatusCode } from '@wundergraph/cosmo-connect/dist/common/common_pb';
 import { GetRoutersRequest, GetRoutersResponse, Router } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_pb';
 import { validate as validateUUID } from 'uuid';
-import { GraphCompositionDTO } from '../../../types/index.js';
+import { GraphCompositionDTO, PlainMessage } from '../../../types/index.js';
 import { FederatedGraphRepository } from '../../repositories/FederatedGraphRepository.js';
 import { GraphCompositionRepository } from '../../repositories/GraphCompositionRepository.js';
 import { RouterMetricsRepository } from '../../repositories/analytics/RouterMetricsRepository.js';