+++ controlplane/src/core/bufservices/federated-graph/migrateFromApollo.ts
@@ -1,4 +1,3 @@
-import { PlainMessage } from '@bufbuild/protobuf';
 import { HandlerContext } from '@connectrpc/connect';
 import { EnumStatusCode } from '@wundergraph/cosmo-connect/dist/common/common_pb';
 import { OrganizationEventName, PlatformEventName } from '@wundergraph/cosmo-connect/dist/notifications/events_pb';
@@ -6,8 +5,9 @@ import {
   MigrateFromApolloRequest,
   MigrateFromApolloResponse,
 } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_pb';
-import { GraphApiKeyJwtPayload } from '../../../types/index.js';
+import { GraphApiKeyJwtPayload, PlainMessage } from '../../../types/index.js';
 import { audiences, signJwtHS256 } from '../../crypto/jwt.js';
+import { UnauthorizedError } from '../../errors/errors.js';
 import { AuditLogRepository } from '../../repositories/AuditLogRepository.js';
 import { FederatedGraphRepository } from '../../repositories/FederatedGraphRepository.js';
 import { DefaultNamespace, NamespaceRepository } from '../../repositories/NamespaceRepository.js';
@@ -16,10 +16,9 @@ import { SubgraphRepository } from '../../repositories/SubgraphRepository.js';
 import { UserRepository } from '../../repositories/UserRepository.js';
 import type { RouterOptions } from '../../routes.js';
 import ApolloMigrator from '../../services/ApolloMigrator.js';
+import { CompositionService } from '../../services/CompositionService.js';
 import { enrichLogger, getLogger, handleError } from '../../util.js';
 import { OrganizationWebhookService } from '../../webhooks/OrganizationWebhookService.js';
-import { UnauthorizedError } from '../../errors/errors.js';
-import { CompositionService } from '../../services/CompositionService.js';
 
 export function migrateFromApollo(
   opts: RouterOptions,