+++ controlplane/src/core/bufservices/playground/getPlaygroundScripts.ts
@@ -1,4 +1,3 @@
-import { PlainMessage } from '@bufbuild/protobuf';
 import { HandlerContext } from '@connectrpc/connect';
 import { EnumStatusCode } from '@wundergraph/cosmo-connect/dist/common/common_pb';
 import {
@@ -8,6 +7,7 @@ import {
 import { PlaygroundScriptsRepository } from '../../repositories/PlaygroundScriptsRepository.js';
 import type { RouterOptions } from '../../routes.js';
 import { enrichLogger, getLogger, handleError } from '../../util.js';
+import type { PlainMessage } from '../../../types/index.js';
 
 export function getPlaygroundScripts(
   opts: RouterOptions,