+++ controlplane/src/core/bufservices/workspace/getWorkspace.ts
@@ -1,10 +1,10 @@
-import { PlainMessage } from '@bufbuild/protobuf';
 import { HandlerContext } from '@connectrpc/connect';
 import { EnumStatusCode } from '@wundergraph/cosmo-connect/dist/common/common_pb';
 import { GetWorkspaceRequest, GetWorkspaceResponse } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_pb';
+import type { PlainMessage } from '../../../types/index.js';
 import type { RouterOptions } from '../../routes.js';
-import { enrichLogger, getLogger, handleError } from '../../util.js';
 import { WorkspaceService } from '../../services/WorkspaceService.js';
+import { enrichLogger, getLogger, handleError } from '../../util.js';
 
 export function getWorkspace(
   opts: RouterOptions,