+++ shared/CHANGELOG.md
@@ -4,6 +4,12 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.49.0](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.48.5...@wundergraph/cosmo-shared@0.49.0) (2026-07-06)
+
+### Features
+
+* upgrade connect v1 to v2 ([#2705](https://github.com/wundergraph/cosmo/issues/2705)) ([c7203f4](https://github.com/wundergraph/cosmo/commit/c7203f47529f75675ca04e6ac4018efe55d4e498)) (@JivusAyrus)
+
 ## [0.48.5](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.48.4...@wundergraph/cosmo-shared@0.48.5) (2026-07-02)
 
 **Note:** Version bump only for package @wundergraph/cosmo-shared